requirements.txt
supabase==2.7.1
python-telegram-bot==20.3
web3==6.15.1
solana==0.29.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.3
aioredis==2.0.1
redis==5.0.1
pytz==2024.1
python-socketio==5.11.2
eth-account==0.11.0
trx-utils==1.0.4
py-algorand-sdk==2.4.0